Changed how collisions between entry colliders are accounted for, now a room will provide the number of all collisions occured with other entry colliders (Previously this was binary, either it was connected with something or not), providing a more accurate connectivity measurement for each level.
Fixed issue with the pivot point of the hallway level piece, re-added to level generator.

Created Graph and GraphNode classes, the purpose of these classes are to be used for translating each Level generated, into a graph, and using some graph theory to calculate the number of connected components in a level (Custom DFS/BFS algorithm).

Created GraphEditor script, that allows for visualizing the connections in a graph.

Implemented Depth First Search algorithm, the number of connected components in a level can now be successfully calculated, the number of connected components in a level is now used as the fitness function.

Found spike trap asset online, integrated into Unity project, added to level generation algorithm to see how it is placed into the maps, right now each level piece has a set of positions in which a trap could potentially be placed, the level generator places as many traps as there are level pieces, randomly in each available position.